New directories distribution. Arduino Robot US, IR and Engines support
