+++ examples/ArduinoRobot/StopnGo.py
@@ -2,11 +2,14 @@ import HALduino.halduino as halduino
 
 def set_engine(direction: int):
     if direction == 0:
-        halduino.setVEngine1(-255, -255)
+        halduino.setSpeedEngines(0, 0)
         print('STOP!')
     elif direction == 1:
-        halduino.setVEngine1(255, 255)
+        halduino.setSpeedEngines(100, 100)
         print('Forward')
 
 def loop():
-    set_engine(1)
+    if halduino.getUS() < 30:
+        set_engine(0)
+    else:
+        set_engine(1)
